style info div
